feat: add Spinner component and integrate it into Button for loading state
